# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/portal_signup_policies:
    get:
      tags:
        - PortalSignupPolicies
      summary: List PortalSignupPolicies
      description: Get a paginated list of PortalSignupPolicies with optional filtering and sorting
      operationId: listPortalSignupPolicies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalSignupPolicyList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PortalSignupPolicies
      summary: Create PortalSignupPolicy
      description: Create a new PortalSignupPolicy
      operationId: createPortalSignupPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortalSignupPolicyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalSignupPolicy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/portal_tokens:
    get:
      tags:
        - PortalTokens
      summary: List PortalTokens
      description: Get a paginated list of PortalTokens with optional filtering and sorting
      operationId: listPortalTokens
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalTokenList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PortalTokens
      summary: Create PortalToken
      description: Create a new PortalToken
      operationId: createPortalToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortalTokenRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalToken'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PortalSignupPolicy:
      type: object
      required:
        - id
        - enabled
        - metadata
      properties:
        id:
          type: string
          format: uuid
        enabled:
          type: boolean
        note:
          nullable: true
          type: string
          maxLength: 200
        updated_by:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreatePortalSignupPolicyRequest:
      type: object
      required:
        - enabled
        - metadata
      properties:
        enabled:
          type: boolean
        note:
          nullable: true
          type: string
          maxLength: 200
        updated_by:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    PortalSignupPolicyList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PortalSignupPolicy'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PortalToken:
      type: object
      required:
        - id
        - user_id
        - token_nonce
        - token_expires_at
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        token_nonce:
          type: string
          maxLength: 64
        token_expires_at:
          type: string
          format: date-time
        status:
          type: string
        rotated_at:
          nullable: true
          type: string
          format: date-time
        rotated_to:
          nullable: true
          type: string
          format: uuid
        revoked_at:
          nullable: true
          type: string
          format: date-time
        revocation_reason:
          nullable: true
          type: string
          maxLength: 120
        last_used_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreatePortalTokenRequest:
      type: object
      required:
        - user_id
        - token_nonce
        - token_expires_at
        - status
        - metadata
      properties:
        user_id:
          type: string
          format: uuid
        token_nonce:
          type: string
          maxLength: 64
        token_expires_at:
          type: string
          format: date-time
        status:
          type: string
        rotated_at:
          nullable: true
          type: string
          format: date-time
        rotated_to:
          nullable: true
          type: string
          format: uuid
        revoked_at:
          nullable: true
          type: string
          format: date-time
        revocation_reason:
          nullable: true
          type: string
          maxLength: 120
        last_used_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    PortalTokenList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PortalToken'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []